+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 
 * Add IE11 compatibility for the `Search` component.
+* Update `BetaFlag` component styling. [#207](https://github.com/mapbox/dr-ui/pull/207)
 
 ## 0.24.0