+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## master
+
+* Add IE11 compatibility for the `Search` component.
+* Update `BetaFlag` component styling. [#207](https://github.com/mapbox/dr-ui/pull/207)
+
 ## 0.24.0
 
 * Add Sentry to `Feedback` component for text feedback issue management. [#198](https://github.com/mapbox/dr-ui/pull/198)